Allow using pagesdisplay in user and file blueprints
The native `pages` section can only be used on page and site blueprints, not on user or file blueprints. As far as I can tell, there is no reason to prevent using `pagesdisplay` everywhere, i.e on a user view to show their latest contributions. But in this case a query value needs to be present. 
https://github.com/getkirby/kirby/blob/3.8.3/config/sections/pages.php#L54